requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
pymongo==4.10.0
